Made Metadata public again - Added ApiStatus.Internal to all internal packages - Repackaged some classes
