+++ chasm/src/main/java/org/quiltmc/chasm/api/ChasmProcessor.java
@@ -11,12 +11,12 @@ import org.quiltmc.chasm.api.tree.MapNode;
 import org.quiltmc.chasm.api.tree.Node;
 import org.quiltmc.chasm.api.util.SuperClassProvider;
 import org.quiltmc.chasm.internal.ChasmSuperClassProvider;
-import org.quiltmc.chasm.internal.LazyClassNode;
 import org.quiltmc.chasm.internal.TransformationApplier;
 import org.quiltmc.chasm.internal.TransformationSorter;
 import org.quiltmc.chasm.internal.TransformerSorter;
 import org.quiltmc.chasm.internal.asm.ChasmClassWriter;
 import org.quiltmc.chasm.internal.metadata.PathMetadata;
+import org.quiltmc.chasm.internal.tree.LazyClassNode;
 import org.quiltmc.chasm.internal.tree.reader.ClassNodeReader;
 import org.quiltmc.chasm.internal.util.PathInitializer;
 import org.slf4j.Logger;
