+++ chasm/src/main/java/org/quiltmc/chasm/internal/asm/package-info.java
@@ -1,0 +1,4 @@
+@ApiStatus.Internal
+package org.quiltmc.chasm.internal.asm;
+
+import org.jetbrains.annotations.ApiStatus;
