+++ chasm/src/main/java/org/quiltmc/chasm/internal/asm/visitor/package-info.java
@@ -1,0 +1,4 @@
+@ApiStatus.Internal
+package org.quiltmc.chasm.internal.asm.visitor;
+
+import org.jetbrains.annotations.ApiStatus;
