--- chasm/src/main/java/org/quiltmc/chasm/internal/metadata/Metadata.java
@@ -1,5 +1,0 @@
-package org.quiltmc.chasm.internal.metadata;
-
-public interface Metadata {
-    Metadata copy();
-}
