+++ chasm/src/main/java/org/quiltmc/chasm/internal/metadata/OriginMetadata.java
@@ -1,6 +1,7 @@
 package org.quiltmc.chasm.internal.metadata;
 
 import org.quiltmc.chasm.api.Transformation;
+import org.quiltmc.chasm.api.metadata.Metadata;
 
 public class OriginMetadata implements Metadata {
     private final String transformerId;
