+++ chasm/src/main/java/org/quiltmc/chasm/internal/package-info.java
@@ -1,0 +1,4 @@
+@ApiStatus.Internal
+package org.quiltmc.chasm.internal;
+
+import org.jetbrains.annotations.ApiStatus;
