+++ chasm/src/main/java/org/quiltmc/chasm/internal/tree/package-info.java
@@ -1,0 +1,4 @@
+@ApiStatus.Internal
+package org.quiltmc.chasm.internal.tree;
+
+import org.jetbrains.annotations.ApiStatus;
