+++ chasm/src/main/java/org/quiltmc/chasm/internal/tree/reader/ClassNodeReader.java
@@ -6,7 +6,7 @@ import org.quiltmc.chasm.api.tree.ListNode;
 import org.quiltmc.chasm.api.tree.MapNode;
 import org.quiltmc.chasm.api.tree.Node;
 import org.quiltmc.chasm.api.tree.ValueNode;
-import org.quiltmc.chasm.internal.LazyClassNode;
+import org.quiltmc.chasm.internal.tree.LazyClassNode;
 import org.quiltmc.chasm.internal.util.NodeConstants;
 
 public class ClassNodeReader {
